Major rework to try and solve circular references
