+++ pycbc/fft/backend_support.py
@@ -71,11 +71,13 @@ def get_backend():
 
 for scheme_name in ["cpu", "mkl", "cuda", "opencl"]:
     try:
-        mod = __import__('backend_' + scheme_name, fromlist = ['.'])
+        mod = __import__('pycbc.fft.backend_' + scheme_name, fromlist = ['_alist', '_adict'])
         _alist = getattr(mod, "_alist")
         _adict = getattr(mod, "_adict")
+        print _alist, _adict
         _update_global_available(_alist, _adict, _all_backends_list,
                                  _all_backends_dict)
-    except ImportError:
+    #except ImportError:
+    #    pass
+    except RuntimeError:
         pass
-
